made code more modulur with header file for the bfs, now when new algorithm needs to be applied a new header file will be included in the source code
