+++ turtlebot3/src/bfs.cpp
@@ -14,6 +14,7 @@
 #include "turtlebot3/Points.h"
 #include "bits/stdc++.h"
 #include "algorithm"
+#include "bfs.h"
 
 // #################################################### //
 
@@ -153,7 +154,7 @@ void BFS_stack_builder(int i, int  j){
   if(visited[i+1][j] == 0){  
 
     if( i+1 < rows ){ // Seeing whether the index is even feasible or no
-      layer.push_back({{i,j},{i+1,j}}); // Parent is going to be its current node
+      layer.push_back({{i,j},{i+1,j}}); // Parent is go ing to be its current node
       visited[i+1][j] = 1;
     }
 
@@ -242,11 +243,11 @@ std::pair<int,int> build_a_tree( int layer_element_count, int flag, std::pair<in
 */   
 void build_path_from_stack(std::pair<int,int> final_index, std::pair<int,int> indx_parent){
 
-  if(path_pub_count == 0){ // Cheching whether the path has been previously built or not
+  if(path_pub_count >= 0){ // Cheching whether the path has been previously built or not
 
     path.push_back(final_index); // Pushing the Final Index
     stack_layer.pop();
-
+    path.clear();
     std::cout<< "Final index was found\n"; // Just to know whether we have been successful or not
     std::vector<std::pair<std::pair<int,int>, std::pair<int,int>>> templayer; // This layer just stores the top of the stack elements
 
@@ -267,118 +268,93 @@ void build_path_from_stack(std::pair<int,int> final_index, std::pair<int,int> in
 
     }
   }
-
+  layer.clear();
   path_pub_count++;
 
-  for(auto &xt: path){
-    std::cout<<xt.first<<" "<<xt.second<<std::endl;
-  }
 }
 
+/*
+  Function      :   Make a Plan 
+  Arguments     :   start index, end index, width, height, resolution
+  Functionality :   Implement BFS
+  Returns       :   Path
+*/
 
-// ##################################################################################################### //
-// #####################################     Main Function    ########################################## //
-// ##################################################################################################### //
+turtlebot3::Points bfs(int start_index, int end_index, int width, int height, float resolution, std::vector<int> costmap){
 
-int main(int argc, char **argv) {
+  ROS_INFO("+-----------------+ +++++++ +-------------------+\n");
+  ROS_INFO("+-----------------+ Running +-------------------+\n");
+  ROS_INFO("+-----------------+ +++++++ +-------------------+\n");
+
+  std::pair<int,int> initial_index, final_index; // pixels of the current grid cell
+
+
+  final_index.first    = (  end_index / width  ); // Need to verify this
+  final_index.second   = (  end_index % height );
+
+  initial_index.first  = ( start_index / width );
+  initial_index.second = ( start_index % width );
+
+  ROS_INFO("end : %d   , end : %d\n", final_index.first, final_index.second);
+  ROS_INFO("start : %d , start : %d\n", initial_index.first, initial_index.second);
   
-  // ########################## Creating Ros nodes, Subsciber, and Publishers ########################## //
-  ros::init(argc, argv, "map_interpretor");
-  ros::NodeHandle n;      // Creating a nodehandler to make relevant subscribers and publishers
+  get_2d_map(data1); //got the 2d map to operate upon
 
-  ros::Subscriber subscriber = n.subscribe("map",1000, currentMap);   // Subscribing to map topic
-  ros::Subscriber odo_sub = n.subscribe("odom", 1000, get_position);  // Subscribing to odom topic
+  turtlebot3::Points P_msg; // This is going to be the final published path
+  geometry_msgs::Point  point_pub;
 
-  ros::Publisher path_publisher = n.advertise<turtlebot3::Points>("path_sub", 1000); // Creating a Publishing node for the path
-  ros::Rate rate(1);   // Defining a rate for overall ros operations
-  int count = 0;
+  if(odom_sub_count > 0){ // Just checking if the subscription is done or not
 
-  // Getting the target from the user
-  std::cout<<"Please Enter the Location in RVIZ and put the orientation here : \n"<<std::endl;
-  std::cin>>des_x>>des_y;
-  std::cin>>des_theta;
+    visited[initial_index.first][initial_index.second] = 1; // Marking the Current node as visited
 
-  tar_sub_count++;
+    layer.push_back({{-1,-1},{initial_index.first,initial_index.second}});  // Initialzing layer0 with the First node later the elements will be added into it
+    stack_layer.push(layer);  // Now, Stack is a collection of all layers till the final index
 
-  while (ros::ok())
-  {
+    int flag = layer.size()-1; // Flag is the signal that all the possible children in the current array has been added and now u can move on to the next layer
 
-    get_2d_map(data1); //got the 2d map to operate upon
-    std::pair<int,int> initial_index, final_index; // pixels of the current grid cell
-    turtlebot3::Points P_msg; // This is going to be the final published path
-    geometry_msgs::Point  point_pub;
+    int layer_element_count = 0; // this will check if all the elements in the layer has been traversed or not
 
-    if(odom_sub_count > 0 and tar_sub_count > 0){ // Just checking if the subscription is done or not
+    // ############################################################################################### //
+    std::pair<int,int> indx_parent;
+    indx_parent = build_a_tree(flag, layer_element_count, final_index); // This will give us the parent of the final node 
 
-      initial_index = distance_to_pixel(curr_x, curr_y);
-      final_index = distance_to_pixel(des_x, des_y);
-      ROS_INFO("START x %d START y %d, END x %d END y %d\n", initial_index.first, initial_index.second, final_index.first, final_index.second);
-      visited[initial_index.first][initial_index.second] = 1; // Marking the Current node as visited
+    if(final_index_found){
 
-      layer.push_back({{-1,-1},{initial_index.first,initial_index.second}});  // Initialzing layer0 with the First node later the elements will be added into it
-      stack_layer.push(layer);  // Now, Stack is a collection of all layers till the final index
+      build_path_from_stack(final_index, indx_parent);
 
-      int flag = layer.size()-1; // Flag is the signal that all the possible children in the current array has been added and now u can move on to the next layer
-
-      int layer_element_count = 0; // this will check if all the elements in the layer has been traversed or not
-
-      // ############################################################################################### //
-      std::pair<int,int> indx_parent;
-      indx_parent = build_a_tree(flag, layer_element_count, final_index); // This will give us the parent of the final node 
-
-      if(final_index_found){
-
-        build_path_from_stack(final_index, indx_parent);
-
-      } 
-      else{
-        std::cout<<"No path Exists"<<std::endl;
-        break;
-      }
-      // Now as there is some path existing 
-
+    } 
+    else{
+      return P_msg;
     }
+    // Now as there is some path existing 
 
-    // ############### Converting the path in terms of the message that we are going to publish ############# //
-    int x_t = path.size();
-    std::cout<<"flag: "<<x_t<<std::endl;
-    geometry_msgs::Point path_pub[x_t];
-
-    for(int i = 0; i < path.size() ; i++){
-      std::pair<double,double> path_temp;
-      path_temp = pixel_to_distance(path[i].first, path[i].second);
-      
-      point_pub.x = path_temp.first;
-      point_pub.y = path_temp.second;
-      path_pub[i] = point_pub; 
-    }
-
-    P_msg.points.clear();
-    P_msg.end_index = des_theta;
-    int index_z = 0;
-    // Now declaring the messages with the help of the points
-    for (auto &it : path_pub) {
-        geometry_msgs::Point point;
-        point.x = (it).x;
-        point.y = (it).y;
-        point.z = 0;
-        P_msg.points.push_back(point);
-        index_z++;
-    }
-    for(int j = 0; j<path.size() ; j++){
-        std::cout<<"x : "<<P_msg.points[j].x<<" y : "<<P_msg.points[j].y<<std::endl;
-    }
-  
-
-    path_publisher.publish(P_msg);
-    ros::spinOnce();
-
-    rate.sleep();
-    ++count;
   }
-  return 0;
-}
 
-// ############################################################################################## //
-// #######################################  The End  ############################################ //
-// ############################################################################################## //
+  // ############### Converting the path in terms of the message that we are going to publish ############# //
+  int x_t = path.size();
+
+  geometry_msgs::Point path_pub[x_t];
+
+  for(int i = 0; i < path.size() ; i++){
+    std::pair<double,double> path_temp;
+    path_temp = pixel_to_distance(path[i].first, path[i].second);
+    
+    point_pub.x = path_temp.first;
+    point_pub.y = path_temp.second;
+    path_pub[i] = point_pub; 
+  }
+
+  P_msg.points.clear();
+  P_msg.end_index = des_theta;
+  int index_z = 0;
+  // Now declaring the messages with the help of the points
+  for (auto &it : path_pub) {
+      geometry_msgs::Point point;
+      point.x = (it).x;
+      point.y = (it).y;
+      point.z = 0;
+      P_msg.points.push_back(point);
+      index_z++;
+  }
+  return P_msg;
+}
